requirements.txt
asgiref==3.2.5
Django==3.0.4
django-countries==6.1.2
django-phone-field==1.8.0
djangorestframework==3.11.0
pip==20.0.2
pytz==2019.3
setuptools==41.2.0
sqlparse==0.3.1